+++ static/zwolle/api/v1/resources.php
@@ -124,6 +124,7 @@ $app->patch('/resources/:resourceType/:resourceId(/:ifcPath+)', function ($resou
 					, 'invariantRulesHold'	=> $session->database->getInvariantRulesHold()
 					, 'requestType'			=> $session->database->getRequestType()
 					, 'sessionRefreshAdvice' => $session->getSessionRefreshAdvice()
+					, 'navTo'				=> $session->getNavToResponse()
 					);
 	
 	print json_encode($result, JSON_PRETTY_PRINT | JSON_UNESCAPED_SLASHES);
@@ -151,6 +152,7 @@ $app->post('/resources/:resourceType/:resourceId/:ifcPath+', function ($resource
 					, 'invariantRulesHold'	=> $session->database->getInvariantRulesHold()
 					, 'requestType'			=> $session->database->getRequestType()
 					, 'sessionRefreshAdvice' => $session->getSessionRefreshAdvice()
+					, 'navTo'				=> $session->getNavToResponse()
 					);
 
 	print json_encode($result, JSON_PRETTY_PRINT | JSON_UNESCAPED_SLASHES);
@@ -176,6 +178,7 @@ $app->delete('/resources/:resourceType/:resourceId/:ifcPath+', function ($resour
 					, 'invariantRulesHold'	=> $session->database->getInvariantRulesHold()
 					, 'requestType'			=> $session->database->getRequestType()
 					, 'sessionRefreshAdvice' => $session->getSessionRefreshAdvice()
+					, 'navTo'				=> $session->getNavToResponse()
 					);
 
 	print json_encode($result, JSON_PRETTY_PRINT | JSON_UNESCAPED_SLASHES);